rename: relocate files to folders
