+++ src/App.js
@@ -1,14 +1,14 @@
-const Lotto = require("./Lotto");
-const Purchase = require("./Purchase");
+const Lotto = require("./service/Lotto");
+const Purchase = require("./service/Purchase");
 const { Console } = require("@woowacourse/mission-utils");
-const { GAME_MESSAGES } = require("./constants");
+const { GAME_MESSAGES } = require("./constants/constants");
 
 class App {
   constructor() {}
   play() {
     Console.readLine(GAME_MESSAGES.ASK_TO_PAY, (cost) => {
       const lottoNumbers = new Purchase(cost);
-      new Lotto(cost, lottoNumbers);
+      const lottoResult = new Lotto(cost, lottoNumbers);
     });
   }
 }
